+++ GeosCore/flexgrid_read_mod.F90
@@ -112,7 +112,7 @@ CONTAINS
     CALL Get_Met_2D( Input_Opt, State_Grid, Q, TRIM(v_name) )
     State_Met%FRLAKE = Q
 
-    ! Read FRLAND
+    ! Read FRLAND (land without lake or ice)
     v_name = "FRLAND"
     CALL Get_Met_2D( Input_Opt, State_Grid, Q, TRIM(v_name) )
     State_Met%FRLAND = Q
@@ -120,7 +120,7 @@ CONTAINS
     ! Read FRLANDIC
     v_name = "FRLANDIC"
     CALL Get_Met_2D( Input_Opt, State_Grid, Q, TRIM(v_name) )
-    State_Met%FRLANDIC = Q
+    State_Met%FRLANDICE = Q
 
     ! Read FROCEAN
     v_name = "FROCEAN"
@@ -217,6 +217,7 @@ CONTAINS
 !
     ! Scalars
     INTEGER            :: t_index            ! Time index
+    INTEGER            :: I, J
     CHARACTER(LEN=16)  :: stamp              ! Time and date stamp
     CHARACTER(LEN=255) :: v_name             ! netCDF variable name
     CHARACTER(LEN=255) :: errMsg             ! Error message
@@ -292,7 +293,18 @@ CONTAINS
     ! Read FRSNO
     v_name = "FRSNO"
     CALL Get_Met_2D( Input_Opt, State_Grid, Q, TRIM(v_name), t_index=t_index )
-    State_Met%FRSNO = Q
+
+    ! The import FRSNO is fraction of land with snow cover. Convert to
+    ! fraction of grid box with snow cover for storage in State_Met.
+    !$OMP PARALLEL DO       &
+    !$OMP DEFAULT( SHARED  )&
+    !$OMP PRIVATE( I, J    )
+    DO J = 1, State_Grid%NY
+    DO I = 1, State_Grid%NX
+       State_Met%FRSNOW(I,J) = Q(I,J) * State_Met%FRLAND(I,J)
+    ENDDO
+    ENDDO
+    !$OMP END PARALLEL DO
 
     !--------------------------------------------------------------------------
     ! For now, skip reading GRN. It's not used in GEOS-Chem. (mps, 9/14/17)